backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis and Caching
redis==5.0.1
celery==5.3.4
flower==2.0.1

# OCR Engines
paddleocr==2.7.3
paddlepaddle==2.6.0
easyocr==1.7.1
pytesseract==0.3.10

# Image Processing
Pillow==10.1.0
opencv-python-headless==4.8.1.78
pdf2image==1.16.3
pypdf==3.17.4
pdfplumber==0.10.3

# Email Integration
aioimaplib==1.0.1
aiosmtplib==3.0.1
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Microsoft Graph (Outlook/Exchange)
msal==1.26.0

# WhatsApp
aiohttp==3.9.1
websockets==12.0

# Accounting Integrations
python-quickbooks==0.9.5
xero-python==5.3.0
requests==2.31.0
httpx==0.25.2

# File Storage
boto3==1.34.7  # S3/DigitalOcean Spaces
minio==7.2.2   # Self-hosted S3-compatible

# ML and AI for Categorization
transformers==4.36.2
sentence-transformers==2.2.2
scikit-learn==1.3.2
torch==2.1.2
numpy==1.26.2

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
python-magic==0.4.27
cryptography==41.0.7
bcrypt==4.1.2
python-dateutil==2.8.2
pytz==2023.3.post1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Monitoring
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.39.1

# Invoice Parsing
invoice2data==0.4.3
